hw_1 fix
