+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/GenerateCategory.java
@@ -4,11 +4,10 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface GenerateCategory {
-
     String category();
     String username();
-
 }
