+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/po/pc/SpendingTable.java
@@ -11,8 +11,21 @@ import static com.codeborne.selenide.Selenide.$;
 public class SpendingTable {
 
     private final ElementsCollection rowsWithSpending = $(".spendings-table tbody").$$("tr");
+    private final SelenideElement deleteSpendingButton = $(".spendings__bulk-actions button");
     private final By cellSelector = By.cssSelector("td");
 
+    public void clickDeleteSpending() {
+        deleteSpendingButton.click();
+    }
+
+    public SpendingTable clickCheckBoxByDescription(String description) {
+        getCheckBox(description).scrollIntoView(true).click();
+        return this;
+    }
+
+    public void checkSizeTable(int size) {
+        rowsWithSpending.shouldHave(size(size));
+    }
 
     private SelenideElement getRowByDescription(String description) {
         return rowsWithSpending.find(text(description));
@@ -21,14 +34,4 @@ public class SpendingTable {
     private SelenideElement getCheckBox(String description) {
         return getRowByDescription(description).$$(cellSelector).first();
     }
-
-    public void clickCheckBoxByDescription(String description) {
-        getCheckBox(description).scrollIntoView(true).click();
-    }
-
-    public void checkSizeTable(int size) {
-        rowsWithSpending.shouldHave(size(size));
-    }
-
-
 }
